requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart

##-e .